CalculateProbabilityHist and Testing Update
General:
-updated Changelog to document these updates

Utils:
-Updated calculateProbabilityHist functions to filter input data vectors of any non finite numbers
-Fixed rare bug in calculateProbabilityHist function where the index was being calculated wrong due to floating point rounding error
-Updated vector_which_median function to cast the return of the distance function to int to prevent compiler warning
-Updated createGaussianDOSVector function by refactoring the code and using the generate function to fill the ouput data vector

tests:
-Added tests to CalculateProbabilityHistTests to check that the data filtering for non finite values is working properly
-Updated CalculateProbabilityHistTests by specifying the double version of the uniform_real_distribution, increasing the number of entries in the data vectors to increase stastically accuracy and reduce likelihood of test failure, and using the generate function to populate the data vectors
-Updated all uses of count_if function by casting the return to int to prevent compiler warning

+++ src/Simulation.cpp
@@ -164,7 +164,7 @@ namespace KMC_Lattice {
 	}
 
 	int Simulation::getN_events() const {
-		return count_if(event_ptrs.begin(), event_ptrs.end(), [](Event* element) {
+		return (int)count_if(event_ptrs.begin(), event_ptrs.end(), [](Event* element) {
 			return (element != nullptr && element->getExecutionTime() > 0);
 		});
 	}
